Add server to x-pack/plugin and remove @kbn/playground package
