--- packages/kbn-playground/README.md
@@ -1,3 +1,0 @@
-# @kbn/playground
-
-Empty package generated by @kbn/generate
