--- packages/kbn-playground/kibana.jsonc
@@ -1,5 +1,0 @@
-{
-  "type": "shared-browser",
-  "id": "@kbn/playground",
-  "owner": "@elastic/enterprise-search-frontend"
-}
