--- packages/kbn-playground/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "@kbn/playground",
-  "private": true,
-  "version": "1.0.0",
-  "license": "SSPL-1.0 OR Elastic License 2.0"
-}
