--- packages/kbn-playground/tsconfig.json
@@ -1,31 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "target/types",
-    "types": [
-      "jest",
-      "node",
-      "react"
-    ]
-  },
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-  ],
-  "exclude": [
-    "target/**/*"
-  ],
-  "kbn_references": [
-    "@kbn/i18n",
-    "@kbn/core",
-    "@kbn/kibana-react-plugin",
-    "@kbn/user-profile-components",
-    "@kbn/i18n-react",
-    "@kbn/config-schema",
-    "@kbn/ml-response-stream",
-    "@kbn/logging",
-    "@kbn/core-elasticsearch-server",
-    "@kbn/security-plugin",
-    "@kbn/core-http-browser",
-  ]
-}
