+++ x-pack/plugins/search_playground/server/index.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { PluginInitializerContext } from '@kbn/core/server';
+
+export async function plugin(initializerContext: PluginInitializerContext) {
+  const { SearchPlaygroundPlugin } = await import('./plugin');
+  return new SearchPlaygroundPlugin(initializerContext);
+}
+
+export type { SearchPlaygroundPluginSetup, SearchPlaygroundPluginStart } from './types';
