+++ x-pack/plugins/search_playground/server/utils/fetch_query_source_fields.ts
@@ -1,15 +1,14 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
 import { FieldCapsResponse, SearchHit } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import { IScopedClusterClient } from '@kbn/core-elasticsearch-server';
 import { get } from 'lodash';
-import { IndicesQuerySourceFields } from '../../types';
+import { IndicesQuerySourceFields } from '../types';
 
 export const fetchFields = async (
   client: IScopedClusterClient,
